+++ TheFanGame/ray.cpp
@@ -16,8 +16,8 @@ ray::~ray()
 
 const void ray::castRay(player* player, world* world, const unsigned int& screenWidth, const unsigned int& screenHeight, unsigned int& i, sf::Vector2f& dir)
 {
-    if (this->m_vertices.getVertexCount() != screenWidth + 1)
-        this->m_vertices.resize(screenWidth + 1);
+    if (this->m_vertices.getVertexCount() != static_cast<size_t>(screenWidth) + 1)
+        this->m_vertices.resize(static_cast<size_t>(screenWidth) + 1);
 
     this->sideDist = sf::Vector2f();
     this->step = sf::Vector2i();
@@ -93,8 +93,8 @@ const void ray::castRay(player* player, world* world, const unsigned int& screen
 
     int lineHeight = (int)(screenHeight / this->perpWallDist);
     sf::Vector2i draw;
-    draw.x = (int)(-lineHeight / 2 + screenHeight / 2 * player->angle);
-    draw.y = (int)(lineHeight / 2 + screenHeight / 2 * player->angle);
+    draw.x = (int)(-lineHeight / 2.f + screenHeight / 2.f * player->angle);
+    draw.y = (int)(lineHeight / 2.f + screenHeight / 2.f * player->angle);
 
     this->drawSE = draw;
 }
@@ -111,7 +111,7 @@ const sf::Vector2i& ray::getMapPos() const noexcept { return this->map; }
 
 const float& ray::getDistance() const noexcept { return this->perpWallDist; }
 
-const void ray::resize(const unsigned int& screenWidth) noexcept { this->m_vertices.resize(screenWidth + 1); }
+const void ray::resize(const unsigned int& screenWidth) noexcept { this->m_vertices.resize(static_cast<size_t>(screenWidth) + 1); }
 
 sf::Vertex& ray::operator[](const std::size_t index)
 {
